Public release 4
This release makes your applications have a prettier and more uniform look in all browsers.
Also, the server is now easier to use. It supports server-side Kalzit code (using the .ks file extension) and index files (like index.ks), which allows similar usecases to PHP. Also, you can implement your own server-side language if you like!
As always, there are many additions to the standard library, and a little more documentation exists as well. Oh, and many things got a lot faster!

But most importantly, it is now much easier to install the project. Just take a look at the README.
Lastly, there is a list of future plans for the first time. So everyone can know there this project is headed.