One year of changes
In the future, the plan is to update Kalzit with multiple much smaller commits - like most other open source projects. This will hopefully make it simpler to track changes - for me and for others. I originally planned to make a release out of this (with a changelog file, new version, etc.), but so many things changed that I lost track of them. Maybe this new strategy works better.