Added more special tree items for common values
This includes calcitSetType and calcitSetAnnotation, as well as "do" and "get" (as long as these are implicitly used by the code generated via the tree generator). If you want to re-define any of these for use in your own scope, you can still do so - but the implicitly generated calls to these functions will always use the original values. This is good, because the other behavior would lead to weird issues when overriding an internally used function accidentally.