Shortened tree item properties
Each tree item has a kind and a value, and each kind of tree item can have a different name for its value. The properties for the most common tree items (strings and variable names) were renamed to "s" and "n", making the resulting files smaller and (potentially) the properties easier to resolve. "kind" is now "k" for the same reasons.
Readability should not suffer too much, because usually there is a constant behind "k" that starts with "KIND_", so we can quickly understand what this is about.