In finished HTML files, the entire app code is now in one place
Before, the code was in two places - one part at the top, with the libraries and runtime definitions, and one part at the bottom, with the actual app code to run. This made the entire thing pretty hard to optimize, so now it is one single blob at the top.