Code blocks are now represented as functions
Before, a code block (something in curly braces {code}) was represented as a syntax tree at runtime. This meant that a function to evaluate these trees always had to be run at runtime - and very often, because code blocks are used all the time in Kalzit. With this new way of representing them, they CAN still use syntax trees to run, but they could also be implemented in a more efficient way without having to re-design the Kalzit runtime.