Attempted to make common routines more GC friendly
The theory is that all values, including those stored in variables, need to be checked at some point to see if they are to be garbage collected - if their values are no longer reachable by any code. So if fewer variables are created or re-defined, this process should, in theory, become slightly faster.