Type system performance improvements
In previous versions, adding a type to a value took a little while, and figuring out if something was a typed value from looking at a compiled tree was not easy. The first problem was solved by simply storing type info in a JavaScript property, which makes all kinds of helper functions unnecessary. The second problem was solved by adding a "typed" kind of tree item, which includes two things: its type, and its value (can only be a code block or a string). This should make the tree easier to evaluate.