Moved the handling of typed variables to the compiler
Previously, the Kalzit runtime was responsible for respecting the type of a given name when using that name to assign a value to it. Now, the Kalzit compiler translated this form "$a ? type = b" into this piece of equivalent code: "$a = type: b". So the runtime requires less logic to function.