Massive update to the equals sign
In previous versions, the equals sign was a normal function that happened to define a variable. The problem was that this was very hard to optimize for performance, while the benefits (you could assign something to a name that was determined at runtime) was practically never used. So in this version, there are some changes:

* The way the equals sign behaves is determined at compile time. If used with a string as the variable, it creates a new variable. If used with anything but a string, it is assumed that you want to change a mutable container. This CAN BREAK SOME OLD CODE!
* Because we can now know which strings are supposed to be variable names, they are unified at compile time. env.resolveName does no longer need to do the check for a valid variable, which should improve performance.
* The equals sign is no longer available as a variable; using it as a variable will cause a runtime error
* env.setInnerWithoutListeners was renamed to qdSet (quick-and-dirty set) to make the name shorter