Changed how functions are compiled
In previous versions, the creation of a function was a function call itself (of "fun"). The problem was that this is hard to optimize for performance, because every aspect of this could potentially have been dynamic (the parameter names, for instance, could in theory have been determined at runtime).

Since this problem existed, and since the potential benefits were never really needed, functions are now compiled in a more efficient, but less dynamic way. The compiler will tell you about the new limitations if you happen to run into one.

There is also a new "noParameterFunctionFromString" function, which you can use to create a function dynamically. Please note that this could get much slower than a "normal" function creation.