Changed the way to create functions from strings
The previous variant would bind the created function to the scope it was created in, which is a) dangerous because you might make unwanted variables accessible, and b) harder to optimize / translate to other languages.

Now, noParameterFunctionFromString is no longer available; instead, there is globalSingleParameterFunctionFromString, which is always run with only the global scope as its parent, and any variables you really need to pass through can be packaged into the single parameter (which you can name however you like).

For an example of how to use this, see the changes in nodejs/files/httpServer.k